【Build Phase】刪除oh-finally-355.mp3，避免Build Fail 【Double Extension】removeZerosFromEnd() 讓 Double Type 1.0 顯示為 1 【醫師建議頁】速率時間顯示時 Int 改 Double 【醫師建議頁】時間的鍵盤改為無小數點的鍵盤 【smartCoachListViewController】更動處待 review

+++ Practice_009/Ad/Models/PatientBorgAndCoach.swift
@@ -1,8 +1,8 @@
 import ObjectMapper
 
 class PatientBorgAndCoach: NSObject {
-    var speed: Int?
-    var time: Int?
+    var speed: Double?
+    var time: Double?
     var preborg: Int?
     var postborg: Int?
     var prebeat: Int?
